# contracts/openapi/longhub-cloud-v1.yaml
openapi: 3.1.0
info:
  title: LongHub Cloud API
  description: 龙枢云台控制面 REST 契约 V1（冻结契约草案）。破坏性变更必须创建新主版本或兼容过渡层。
  version: "1.0.0"
servers:
  - url: https://api.longhub.example.com/v1
...
security:
  - bearerAuth: []
paths:
...
  /admin/model-usage:
    get:
      tags: [model]
      summary: 查询设备与租户的日/月模型计量聚合
      operationId: listAdminModelUsage
      responses:
        "200": { description: 模型 Token、成本和结果聚合 }
        default: { $ref: "#/components/responses/Error" }
  /admin/devices:
    get:
      tags: [identity]
      summary: 列出设备运营状态；不返回设备凭据或指纹
      operationId: listAdminDevices
      responses:
        "200": { description: 设备列表、最后在线/模型成功、错误摘要与策略字段 }
        default: { $ref: "#/components/responses/Error" }
...
  /admin/activation-codes:
    get:
      tags: [identity]
      summary: 管理面列出授权码脱敏记录
      operationId: listActivationCodes
      responses:
        "200": { description: 不含明文与摘要的授权码列表 }
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [identity]
      summary: 管理面生成授权码；明文只在本次响应返回
      operationId: createActivationCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivationCodeRequest"
      responses:
        "201": { description: 授权码已生成 }
        default:
          $ref: "#/components/responses/Error"
  /admin/activation-codes/{activationCodeId}/revoke:
    post:
      tags: [identity]
      summary: 管理面撤销授权码，并立即阻断绑定设备的产品 API
      operationId: revokeActivationCode
      parameters:
        - name: activationCodeId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 授权码已撤销 }
        default:
          $ref: "#/components/responses/Error"
...
  /entitlements:
    get:
      tags: [entitlement]
      summary: 查询当前用户/设备的有效授权
      operationId: listEntitlements
      responses:
        "200":
          description: 授权列表
          content:
            application/json:
              schema:
                type: object
                required: [entitlements]
                properties:
                  entitlements:
                    type: array
                    items:
                      $ref: "#/components/schemas/Entitlement"
        "4XX":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /admin/entitlements:
    post:
      tags: [entitlement]
      summary: 管理面：为设备授予套装授权（Console 使用管理凭据）
      operationId: grantEntitlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id, pack_id]
              properties:
                device_id:
                  type: string
                pack_id:
                  type: string
                expires_at:
                  type: string
                  format: date-time
      responses:
        "201":
          description: 授权已授予
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entitlement"
        "4XX":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /admin/packs:
    post:
      tags: [release]
      summary: 管理面：上传套装，云端签名后发布（Console 使用管理凭据；同版本重发返回 409）
      operationId: publishPack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest, files]
              properties:
                manifest: {}
                files:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackReleaseSummary"
        "4XX":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details: {}
...
    CreateActivationCodeRequest:
      type: object
      properties:
        tenant_id: { type: string, default: tenant-default }
        label: { type: string, maxLength: 200 }
        max_uses: { type: integer, minimum: 1, maximum: 10000, default: 1 }
        expires_in_days: { type: integer, minimum: 1, maximum: 3650, default: 365 }
        pack_ids:
          type: array
          maxItems: 32
          items: { type: string }
...
    Entitlement:
      type: object
      required: [entitlement_id, pack_id, scope, status, expires_at]
      properties:
        entitlement_id:
          type: string
        pack_id:
          type: string
        scope:
          type: string
          enum: [tenant, user, device]
        status:
          type: string
          enum: [active, suspended, revoked]
        expires_at:
          type: string
          format: date-time
...
    PackReleaseSummary:
      type: object
      required: [pack_id, version, status]
      properties:
        pack_id:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [active, revoked]
        digest:
          type: string
        signature_key_id:
          type: string
        created_at:
          type: string
          format: date-time
...